And so the orchestration was split, like Alice through her looking-glass —   one PSA on this side, one AlgaDesk on the other, both reading from the   same WelcomeEmailCopy bundle as if it were the Caterpillar's hookah of   templates. The "AlgaDesk Customer" tag follows the white rabbit into the   Nine Minds tenant; the Stripe checkout drops its base-fee ballast and   sails on per-user prices alone. Curiouser and curiouser. 🐇✉️ 🎫
