+++ .env.example
@@ -139,6 +139,9 @@ STRIPE_LICENSE_PRICE_ID=price_your_price_id_here
 # STRIPE_EARLY_ADOPTERS_USER_PRICE_ID=price_early_adopters_per_user_here
 # STRIPE_EARLY_ADOPTERS_BASE_ANNUAL_PRICE_ID=price_early_adopters_base_annual_here
 # STRIPE_EARLY_ADOPTERS_USER_ANNUAL_PRICE_ID=price_early_adopters_per_user_annual_here
+# AlgaDesk prices (per-user only, no base fee — tenants with product_code='algadesk')
+# STRIPE_ALGADESK_USER_PRICE_ID=price_algadesk_per_user_here
+# STRIPE_ALGADESK_USER_ANNUAL_PRICE_ID=price_algadesk_per_user_annual_here
 
 # Email Configuration
 EMAIL_ENABLE=false  # Required: Boolean
