+++ ee/temporal-workflows/src/workflows/index.ts
@@ -1,4 +1,5 @@
 export * from './tenant-creation-workflow.js';
+export * from './algadesk-tenant-creation-workflow.js';
 export * from './test-email-workflow.js';
 export * from './extension-domain-workflow.js';
 export * from './portal-domains/registration.workflow.js';
